# api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Coletor Patrimonial API",
    "version": "1.0.0",
    "description": "Documentacao OpenAPI da API do Coletor Patrimonial. A autenticacao protegida usa JWT no header `Authorization: Bearer <token>`. O token precisa conter `uid` e `tenantId` validos.",
    "contact": {
      "name": "Coletor Patrimonial Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor local"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT retornado pelos endpoints de autenticacao."
      }
    },
    "parameters": {
...
      "Page": {
        "name": "page",
        "in": "query",
        "required": false,
        "description": "Pagina da listagem.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Quantidade maxima de itens por pagina.",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 1000,
          "default": 50
        }
      },
      "Since": {
        "name": "since",
        "in": "query",
        "required": false,
        "description": "Filtro por data ISO8601 usando `updatedAt >= since`.",
        "schema": {
          "type": "string",
          "format": "date-time"
        }
      },
      "SearchQ": {
        "name": "q",
        "in": "query",
        "required": false,
        "description": "Prefixo de busca.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
      "SearchField": {
        "name": "field",
        "in": "query",
        "required": false,
        "description": "Campo pesquisado. O alias `local` mapeia para `localizacaoNome`.",
        "schema": {
          "type": "string",
          "enum": [
            "codigo",
            "placa",
            "descricao",
            "local"
          ]
        }
      },
      "NrInventario": {
        "name": "nrInventario",
        "in": "query",
        "required": false,
        "description": "Numero do inventario usado para escopo de consulta, catalogos, busca por codigo e exportacao.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      },
...
    },
    "schemas": {
      "ValidationErrorDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "field"
          },
          "value": {
            "nullable": true
          },
          "msg": {
            "type": "string",
            "example": "Senha deve ter pelo menos 6 caracteres"
          },
          "path": {
            "type": "string",
            "example": "password"
          },
          "location": {
            "type": "string",
            "example": "body"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Dados invalidados"
          },
          "message": {
            "type": "string",
            "example": "Verifique os dados enviados"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetail"
            }
          },
          "stack": {
            "type": "string",
            "description": "Disponivel apenas em desenvolvimento."
          }
        }
      },
...
      "InventoryItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "mcf42g8c-a1b2c3"
          },
          "codigo": {
            "type": "string",
            "nullable": true
          },
          "placa": {
            "type": "string",
            "nullable": true
          },
          "descricao": {
            "type": "string",
            "nullable": true
          },
          "localizacaoNome": {
            "type": "string",
            "nullable": true
          },
          "situacaoNome": {
            "type": "string",
            "nullable": true
          },
          "estadoConservacaoNome": {
            "type": "string",
            "nullable": true
          },
          "dsObservacao": {
            "type": "string",
            "nullable": true
          },
          "codigoLocalizacao": {
            "type": "string",
            "nullable": true
          },
          "codigoSituacao": {
            "type": "string",
            "nullable": true
          },
          "codigoEstado": {
            "type": "string",
            "nullable": true
          },
          "nrInventario": {
            "type": "string",
            "nullable": true
          },
          "valorAtual": {
            "type": "number",
            "nullable": true,
            "example": 2499.9
          },
          "inventariadoPor": {
            "type": "string",
            "nullable": true,
            "example": "Marcelo"
          },
          "StatusBem": {
            "type": "string",
            "nullable": true,
            "example": "Bem Inventariado"
          },
          "statusBem": {
            "type": "string",
            "nullable": true,
            "example": "Bem Inventariado"
          },
          "status": {
            "type": "string",
            "nullable": true,
            "example": "Bem Inventariado"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "InventoryPayload": {
        "type": "object",
        "description": "Payload aceito pelo backend. A API aceita tanto os nomes canonicos quanto aliases legados para localizacao, situacao, estado e observacao.",
        "properties": {
          "id": {
            "type": "string",
            "description": "Opcional no sync para conciliacao por ID."
          },
          "codigo": {
            "type": "string"
          },
          "placa": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "localizacaoNome": {
            "type": "string"
          },
          "localizacao": {
            "type": "string"
          },
          "local": {
            "type": "string"
          },
          "situacaoNome": {
            "type": "string"
          },
          "situacao": {
            "type": "string"
          },
          "estadoConservacaoNome": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "dsObservacao": {
            "type": "string",
            "maxLength": 500
          },
          "observacao": {
            "type": "string",
            "maxLength": 500
          },
          "observacoes": {
            "type": "string",
            "maxLength": 500
          },
          "codigoLocalizacao": {
            "type": "string"
          },
          "cdLocalizacao": {
            "type": "string"
          },
          "codigoSituacao": {
            "type": "string"
          },
          "cdSituacao": {
            "type": "string"
          },
          "codigoEstado": {
            "type": "string"
          },
          "cdEstadoConser": {
            "type": "string"
          },
          "nrInventario": {
            "type": "string"
          },
          "valorAtual": {
            "oneOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "example": "2.499,90"
              }
            ]
          },
          "StatusBem": {
            "type": "string"
          },
          "statusBem": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "inventariadoPor": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "codigo": "123456",
          "placa": "000012345678",
          "descricao": "Notebook Dell Latitude",
          "local": "ALMOXARIFADO CENTRAL",
          "situacao": "EM USO",
          "estado": "BOM",
          "observacoes": "Patrimonio localizado na prateleira 3",
          "cdLocalizacao": "0101",
          "cdSituacao": "01",
          "cdEstadoConser": "01",
          "nrInventario": "2026-0001",
          "valorAtual": "2.499,90",
          "statusBem": "Bem Inventariado",
          "inventariadoPor": "Marcelo Dev"
        }
      },
      "InventoryPagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 50
          },
          "total": {
            "type": "integer",
            "example": 182
          },
          "totalPages": {
            "type": "integer",
            "example": 4
          }
        }
      },
      "InventoryListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/InventoryItem"
                }
              },
              "pagination": {
                "$ref": "#/components/schemas/InventoryPagination"
              }
            }
          }
        }
      },
...
      "InventoryMutationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Item atualizado com sucesso"
          },
          "data": {
            "$ref": "#/components/schemas/InventoryItem"
          },
          "skipped": {
            "type": "boolean",
            "description": "Pode aparecer em atualizacoes protegidas por regra de negocio.",
            "example": false
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Falha de validacao do request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Token ausente, invalido, expirado ou sem tenant valido.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Erro interno do servidor.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/inventory": {
      "get": {
        "tags": [
          "Inventory"
        ],
        "summary": "Listar itens do inventario",
        "description": "Retorna uma lista paginada filtrada pelo tenant autenticado. A busca textual usa prefixo (`LIKE q%`).",
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Since"
          },
          {
            "$ref": "#/components/parameters/SearchQ"
          },
          {
            "$ref": "#/components/parameters/SearchField"
          },
          {
            "$ref": "#/components/parameters/NrInventario"
          }
        ],
        "responses": {
          "200": {
            "description": "Itens retornados com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryListResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "Inventory"
        ],
        "summary": "Criar item de inventario",
        "description": "Cria um novo item vinculado ao usuario autenticado. O backend normaliza aliases legados do payload.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryMutationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
  }
}